lams get return types, so can tc recursive ones
The actual way this is done may not be ideal - I extended `tc` with a
`shallow` argument, which currently only affects one thing - whether
lambda bodys are checked. If the argument is true, we just take the
specified type as an assumption. This argument is only set to true when
checking the let binding position, and it is always set to false
everywhere else, which allows `x = () : t { ... mentions x }` to check,
but feels (to me) a little special cased. In particular,
`x = if blad {... mentions x } ...` won't work, but maybe it should?

This also includes an implementation of <each>, though the way it's done
seems pretty clumsy...