tc returns type env, evalTemplate takes envs.
This makes the template evaluation a lot more straightforward - we parse
the program in order to construct a type/value environment that we can
run expressions within. In order to do this, we needed to be able to
extract a type and value env from the 'program' (which is a series of
let bindings with no body). The value environment we could already do,
by simply making the body be a lambda and grabbing it's enviroment. For
the type environment, this changes `tc` to always return the type
environment. This seems a little bit odd, because almost always we want
to discard it, but the one place we do need it, we really need it...